refactor: moudle-api 삭제
- 각 바운더리 컨텍스트 별로 배포 가능하도록 분리
