+++ build.gradle
@@ -8,7 +8,7 @@ repositories {
     mavenCentral()
 }
 
-subprojects {
+allprojects {
     group = 'camp.nextstep.edu'
     version = '0.0.1-SNAPSHOT'
 
@@ -25,10 +25,22 @@ subprojects {
 
     dependencies {
         implementation 'org.springframework.boot:spring-boot-starter-actuator'
+        implementation 'org.springframework.boot:spring-boot-starter-web'
+        implementation 'org.springframework.boot:spring-boot-starter-data-jpa'
+        implementation 'org.flywaydb:flyway-core'
+        runtimeOnly 'com.h2database:h2'
+        testImplementation 'io.rest-assured:rest-assured:3.3.0'
         testImplementation 'org.springframework.boot:spring-boot-starter-test'
     }
 
     test {
         useJUnitPlatform()
     }
+}
+
+
+dependencies {
+    implementation project(':module-common')
+    implementation project(':module-menu')
+    implementation project(':module-order')
 }
