--- module-api/build.gradle
@@ -1,18 +1,0 @@
-bootJar {
-    enabled=true
-}
-
-jar {
-    enabled=true
-}
-
-dependencies {
-    implementation project(':module-menu')
-    implementation project(':module-order')
-
-    implementation 'org.springframework.boot:spring-boot-starter-web'
-    implementation 'org.flywaydb:flyway-core'
-    runtimeOnly 'com.h2database:h2'
-    testImplementation 'io.rest-assured:rest-assured:3.3.0'
-}
-
