+++ module-common/build.gradle
@@ -6,7 +6,4 @@ jar {
     enabled=true
 }
 
-dependencies {
-    implementation 'org.springframework.boot:spring-boot-starter-data-jpa'
-}
