+++ module-common/src/main/java/kitchenpos/common/config/ControllerExceptionHandler.java
@@ -1,4 +1,4 @@
-package kitchenpos.config;
+package kitchenpos.common.config;
 
 import kitchenpos.common.exception.BusinessException;
 import kitchenpos.common.exception.NotFoundException;
