+++ module-menu/build.gradle
@@ -7,11 +7,5 @@ jar {
 }
 
 dependencies {
-    compile project(":module-common")
-
-    implementation 'org.springframework.boot:spring-boot-starter-data-jpa'
-    implementation 'org.flywaydb:flyway-core'
-
-    runtimeOnly 'com.h2database:h2'
+    implementation project(':module-common')
 }
-
