+++ module-menu/src/main/java/kitchenpos/menu/ui/MenuGroupRestController.java
@@ -1,4 +1,4 @@
-package kitchenpos.menu;
+package kitchenpos.menu.ui;
 
 import java.net.URI;
 import java.util.List;
