+++ module-menu/src/main/java/kitchenpos/product/ui/ProductRestController.java
@@ -1,4 +1,4 @@
-package kitchenpos.product;
+package kitchenpos.product.ui;
 
 import java.net.URI;
 import java.util.List;
