+++ module-order/build.gradle
@@ -7,12 +7,7 @@ jar {
 }
 
 dependencies {
-    compile project(":module-common")
-    implementation project(":module-menu")
-
-    implementation 'org.springframework.boot:spring-boot-starter-data-jpa'
-    implementation 'org.flywaydb:flyway-core'
-
-    runtimeOnly 'com.h2database:h2'
+    implementation project(':module-common')
+    implementation project(':module-menu')
 }
