+++ module-order/src/main/java/kitchenpos/common/config/AsyncConfig.java
@@ -1,4 +1,4 @@
-package kitchenpos.config;
+package kitchenpos.common.config;
 
 import java.util.concurrent.Executor;
 import org.springframework.context.annotation.Bean;
