+++ src/main/resources/db/migration/V1__Initialize_project_tables.sql
