+++ src/main/resources/db/migration/V4__modified_project_tables.sql
