+++ src/test/resources/db/migration/V3__updated_project_tables.sql
