+++ src/test/resources/db/migration/V4__modified_project_tables.sql
