+++ src/test/resources/truncate.sql
